requirements.txt
amqp==5.0.5
arrow==1.0.2
asgiref==3.3.1
autopep8==1.5.5
billiard==3.6.3.0
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==3.1.7
djangorestframework==3.12.4
idna==2.10
kombu==5.0.2
numpy==1.19.5
pandas==1.2.3
prompt-toolkit==3.0.16
pycodestyle==2.6.0
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.3
vine==5.0.0
wcwidth==0.2.5
